Refactor assumeAceValues
This commit refactors assumeAceValues to ensure that it is more readable, additionally, the function now changes checks against multiple Ace value cards as opposed to just the first Ace card in the hand. An additional unit test has been added for when more than two Ace cards exist in a hand.